raflow: MOVE fees handler function for pet/vehicles to raflow_pets/vehicles.go
